requirements.txt
torch>=1.12.0
torchvision>=0.13.0
torchdiffeq>=0.2.3
transformers>=4.25.0
clip
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.60.0
pyyaml>=5.4.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
scikit-learn>=1.0.0
